Renamed inference agent
