+++ src/agents/inference_agent/EvolutionAgentNode.h
@@ -1,5 +1,5 @@
 /**
- * @file distributed_inference_control_agent_node.h
+ * @file EvolutionAgentNode.h
  * @brief Distributed inference control agent node
  * is responsible for sending and receiving messages
  * to the distributed inference control agent.
