+++ src/agents/inference_agent/InferenceRequest.cc
@@ -1,4 +1,4 @@
-#include "inference_request.h"
+#include "InferenceRequest.h"
 
 #include <memory>
