+++ src/tests/cpp/inference_agent_test.cc
@@ -1,4 +1,4 @@
-#include "inference_agent.h"
+#include "InferenceAgent.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
